feat: Add react-icons package and implement icons in Navbar component
This commit adds the "react-icons" package as a dependency and imports the necessary icons in the Navbar component. The FaBars and FaTimes icons from the react-icons library are used to represent the open and close icons for the mobile navigation menu. This addition enhances the visual appeal and functionality of the Navbar component.