feat: Update Notfound page with informative message
The Notfound page has been updated to display a more informative message when a user navigates to a non-existent page. Instead of just displaying "Notfound", the page now shows a centered message stating "The page you are looking for does not exist". This change improves the user experience by providing clearer feedback when encountering a 404 error.