feat: Add react-hot-toast package for toast notifications
The "react-hot-toast" package has been added as a dependency to the project. This package provides a simple and customizable way to display toast notifications. By using this package, the application can show success and error messages to the user when submitting forms or performing other actions. This addition enhances the user experience by providing visual feedback for important events.